Revert "Refactor schema factory as an interface (#4)"
This reverts commit b6fd577b6444009136b82f26ee861b226a02983c.
